Extracted TableOfContents into a separate component
ref https://linear.app/ghost/issue/AP-634/table-of-contents-in-reader-view
